Split the jobboard tests to separate files
1. I split the jobboard tests to 3 separate files according to
   categories.
